% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIterations.R
\name{calcIterations}
\alias{calcIterations}
\title{Calculate the number of iterations for the MCMS run}
\usage{
calcIterations(Peptides, round = TRUE)
}
\arguments{
\item{Peptides}{the \code{Peptides}}

\item{round}{whether to round the number up to the next power of 10 (default). If \code{FALSE} the value is only rounded up to the next integer.}
}
\value{
an integer
}
\description{
The calculation is based on the number of parameters (c and o)
}
\examples{
data(ENSTest, var.model)
ENSTestProtein <- Protein(ENSTest)
ENSTestModel <- Peptides(ENSTestProtein)
calcIterations(ENSTestModel)
}
