% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLikelihood.R
\name{getLikelihood}
\alias{getLikelihood}
\title{Compute the likelihood of a Peptides model}
\usage{
getLikelihood(Peptides, var.model, c = Peptides@c, o = Peptides@o,
  c.prior.sd = sqrt(2), o.prior.shape1 = 0.5, o.prior.shape2 = 0.5,
  verbose = FALSE)
}
\arguments{
\item{Peptides}{the Peptides model data}

\item{var.model}{the variance model}

\item{c, o}{values of the parameters for which to estimate the likelihood}

\item{c.prior.sd, o.prior.shape1, o.prior.shape2}{parameters for the prior}

\item{verbose}{if TRUE, will print internal details of the Monte Carlo object (elements and pointers)}
}
\description{
Compute the likelihood of a Peptides model
}
\examples{
data(ENSTest, var.model)
ENSTestProtein <- Protein(ENSTest)
ENSTestModel <- Peptides(ENSTestProtein)
getLikelihood(ENSTestModel, var.model)
# With custom c
someC <- ENSTestModel@c
someC[2] <- 15
getLikelihood(ENSTestModel, var.model, c = someC)
}
