% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.matrices.R
\name{make.sample.dependency.matrix}
\alias{make.sample.dependency.matrix}
\title{The \code{sample.dependency.matrix} defines the dependency between the concentration and samples.
It will solve simple circular dependencies and encode the last link as a function of other samples}
\usage{
make.sample.dependency.matrix(sample, reference)
}
\arguments{
\item{sample, reference}{the aligned vectors of sample names that represent the sample itself and its reference}
}
\description{
The \code{sample.dependency.matrix} defines the dependency between the concentration and samples.
It will solve simple circular dependencies and encode the last link as a function of other samples
}
\examples{
data(ENSTest)
make.sample.dependency.matrix(ENSTest$sample, ENSTest$reference)
}
