% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.matrices.R
\name{make.sites.activation.matrix}
\alias{make.sites.activation.matrix}
\title{The \code{sites.activation} indicates which observed ratios (rows) cover have the given modification (column) active}
\usage{
make.sites.activation.matrix(data, modifications)
}
\arguments{
\item{data}{the peptide data, as for the \code{\link{Protein}} function}

\item{modifications}{the list of unique modifications to be considered}
}
\description{
The \code{sites.activation} indicates which observed ratios (rows) cover have the given modification (column) active
}
\examples{
data(ENSTest)
make.sites.activation.matrix(ENSTest, unique(ENSTest$modifications))
}
