% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.matrices.R
\name{make.sites.coverage.matrix}
\alias{make.sites.coverage.matrix}
\title{The \code{sites.coverage} matrix indicates which observed ratios (rows) cover which modification site (column)}
\usage{
make.sites.coverage.matrix(data, modifications)
}
\arguments{
\item{data}{the peptide data, as for the \code{\link{Protein}} function}

\item{modifications}{the list of modifications to be considered}
}
\description{
The \code{sites.coverage} matrix indicates which observed ratios (rows) cover which modification site (column)
}
\examples{
data(ENSTest)
make.sites.coverage.matrix(ENSTest, unique(ENSTest$modifications))
}
