% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_ms.R
\name{mcms_cv}
\alias{mcms_cv}
\title{This function performs cross-validation to estimate the optimal lambda values and runs the mcnnm_wc model.	
If post_estimation is set, then the estimated parameters are returned additionally in the list.}
\usage{
mcms_cv(
  M,
  X,
  Z,
  B,
  mask,
  num_B_cov,
  to_add_ID = TRUE,
  to_estimate_u = TRUE,
  to_estimate_v = TRUE,
  to_estimate_b = TRUE,
  to_estimate_H = TRUE,
  niter = 100,
  rel_tol = 1e-05,
  to_normalize = TRUE,
  cv_ratio = NA,
  cv_criterion = "mse",
  num_folds = 3,
  n_config = 150,
  cube_search = TRUE,
  is_quiet = TRUE,
  post_estimation = TRUE,
  model_selection_H = TRUE,
  model_selection_b = TRUE,
  seed = NA
)
}
\arguments{
\item{M}{Outcome matrix}

\item{X}{Units covariate matrix (Time-invariant)}

\item{Z}{Time covariate matrix (Unit-invariant)}

\item{B}{Unit-Time covariate matrix}

\item{mask}{Treatment mask matrix (1 if control, 0 if treated)}

\item{num_B_cov}{Number of unit-time covariates}

\item{to_add_ID}{Whether to add an identity matrix to the X and Z matrices. (default: TRUE)}

\item{to_estimate_u}{Whether to estimate the units' fixed effects. (default: TRUE)}

\item{to_estimate_v}{Whether to estimate the time's fixed effects. (default: TRUE)}

\item{to_estimate_b}{Whether to estimate the unit-time varying covariate parameters. (default: TRUE)}

\item{to_estimate_H}{Whether to estimate the unit and time covariates linking parameters. (default: TRUE)}

\item{niter}{Number of iterations for the coordinate descent steps. (default: 100)}

\item{rel_tol}{Relative tolerance for the convergence of the coordinate descent steps. (default: 1e-5)}

\item{to_normalize}{Whether to normalize the X, Z, and B matrices. (default: TRUE)}

\item{cv_ratio}{Share of observations that is used for training in cross-validation. If NA is provided, then the share is equal to the share of non-treated observations in the sample (default: NA)}

\item{cv_criterion}{One of c("mse","1.se","both"). Optimality criterion for cross-validation. "mse" selects the optimal lambda configuration based on mse, "1.se" the largest lambda configuration for which the error is lower than the minimal mse + one standards error at optimal level. "both" returns both estimates. (Default: "mse")}

\item{num_folds}{Number of folds in cross-validation (default: 3)}

\item{n_config}{Number of configurations of lambda that are evaluated in cross-validation (default: 150)}

\item{cube_search}{Search method for lambda configurations. (default: TRUE)
If set to true, then configurations are evaluated in a hypercube fashion with zooming in on best edge every iteration
If set to false, lambda values are spaned on a grid.}

\item{is_quiet}{Whether to print the progress of the algorithm. (default: FALSE)}

\item{post_estimation}{Whether a post regularization estimation without regularization should be estimated. (default: TRUE)}

\item{model_selection_H}{Whether the unit and time varying covariate link matrix should be regularized. (default: TRUE)}

\item{model_selection_b}{Whether the unit-time varying covariates parameters should be regularized. (default: TRUE)}

\item{seed}{Any integer to control the random seed. NA for a random seed (default: NA)}
}
\value{
: List containing the following elements:
	    - "u": Estimated units' fixed effects
		- "v": Estimated time's fixed effects
		- "B": Estimated unit-time covariates' coefficients
		- "H": Estimated H matrix
		- "tau": The estimate treatment effect
		- "lambda_L": Estimated lambda_L value
		- "lambda_H": Estimated lambda_H value
		- "lambda_b": Estimated lambda_b value
}
\description{
This function performs cross-validation to estimate the optimal lambda values and runs the mcnnm_wc model.	
If post_estimation is set, then the estimated parameters are returned additionally in the list.
}
\examples{
mcms_cv(M, X, Z, B, mask, num_B_cov)

}
