% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_ms.R
\name{mcms_lam_range}
\alias{mcms_lam_range}
\title{This function is used to estimate the maximum lambda_L, lambda_H, and lambda_b values for the MCNNM algorithm.}
\usage{
mcms_lam_range(
  M,
  X,
  Z,
  B,
  mask,
  num_B_cov,
  to_add_ID = TRUE,
  to_estimate_u = TRUE,
  to_estimate_v = TRUE,
  niter = 1e-05,
  rel_tol = 1e-05,
  to_normalize = TRUE,
  is_quiet = FALSE
)
}
\arguments{
\item{M}{Outcome matrix}

\item{X}{Units covariate matrix (Time-invariant)}

\item{Z}{Time covariate matrix (Unit-invariant)}

\item{B}{Unit-Time covariate matrix}

\item{mask}{Treatment mask matrix (1 if control, 0 if treated)}

\item{num_B_cov}{Number of unit-time covariates}

\item{to_add_ID}{Whether to add an identity matrix to the X and Z matrices. (default: TRUE)}

\item{to_estimate_u}{Whether to estimate the units' fixed effects. (default: TRUE)}

\item{to_estimate_v}{Whether to estimate the time's fixed effects. (default: TRUE)}

\item{niter}{Number of iterations for the coordinate descent steps. (default: 100)}

\item{rel_tol}{Relative tolerance for the convergence of the coordinate descent steps. (default: 1e-5)}

\item{to_normalize}{Whether to normalize the X, Z, and B matrices. (default: TRUE)}

\item{is_quiet}{Whether to print the progress of the algorithm. (default: FALSE)}
}
\value{
(List) List of the maximum lambda_L, lambda_H, and lambda_b values
}
\description{
This function is used to estimate the maximum lambda_L, lambda_H, and lambda_b values for the MCNNM algorithm.
}
\examples{
mcms_lam_range(M, X, Z, B, mask, num_B_cov)
}
