% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_ms.R
\name{mcms_lambda_analysis}
\alias{mcms_lambda_analysis}
\title{This function estimate deviations from optimal lambda
If post_estimation is set, then the estimated parameters are returned additionally in the list.}
\usage{
mcms_lambda_analysis(
  M,
  mask,
  X = NA,
  Z = NA,
  B = NA,
  num_B_cov = NA,
  to_normalize = TRUE,
  to_estimate_u = TRUE,
  to_estimate_v = TRUE,
  to_estimate_b = TRUE,
  to_estimate_H = TRUE,
  to_add_ID = FALSE,
  niter = 100,
  n_config = 100,
  rel_tol = 1e-05,
  is_quiet = TRUE,
  n_lambda = 50,
  cv_ratio = NA,
  num_folds = 3,
  cube_search = TRUE,
  post_estimation = TRUE,
  impose_null = TRUE,
  model_selection_H = TRUE,
  model_selection_b = TRUE,
  iter_lambda = 100,
  seed = NA,
  file_path = "./"
)
}
\arguments{
\item{M}{Outcome matrix}

\item{mask}{Treatment mask matrix (1 if control, 0 if treated)}

\item{X}{Units covariate matrix (Time-invariant) (default: NA)}

\item{Z}{Time covariate matrix (Unit-invariant) (default: NA)}

\item{B}{Unit-Time covariate matrix (default: NA)}

\item{num_B_cov}{Number of unit-time covariates (default: NA)}

\item{to_normalize}{Whether to normalize the X, Z, and B matrices. (default: TRUE)}

\item{to_estimate_u}{Whether to estimate the units' fixed effects. (default: TRUE)}

\item{to_estimate_v}{Whether to estimate the time's fixed effects. (default: TRUE)}

\item{to_estimate_b}{Whether to estimate the unit-time varying covariate parameters. (default: TRUE)}

\item{to_estimate_H}{Whether to estimate the unit and time covariates linking parameters. (default: TRUE)}

\item{to_add_ID}{Whether to add an identity matrix to the X and Z matrices. (default: TRUE)}

\item{niter}{Number of iterations for the coordinate descent steps. (default: 100)}

\item{n_config}{Number of configurations of lambda that are evaluated in cross-validation (default: 150)}

\item{rel_tol}{Relative tolerance for the convergence of the coordinate descent steps. (default: 1e-5)}

\item{is_quiet}{Whether to print the progress of the algorithm. (default: FALSE)}

\item{cv_ratio}{Share of observations that is used for training in cross-validation. If NA is provided, then the share is equal to the share of non-treated observations in the sample (default: NA)}

\item{num_folds}{Number of folds in cross-validation (default: 3)}

\item{cube_search}{Search method for lambda configurations. (default: TRUE)
If set to true, then configurations are evaluated in a hypercube fashion with zooming in on best edge every iteration
If set to false, lambda values are spaned on a grid.}

\item{post_estimation}{Whether a post regularization estimation without regularization should be estimated. (default: TRUE)}

\item{impose_null}{Whether null-hypothesis should be imposed. Implies that all observations also treated observations are used in model estimation. (default: TRUE)}

\item{model_selection_H}{Whether the unit and time varying covariate link matrix should be regularized. (default: TRUE)}

\item{model_selection_b}{Whether the unit-time varying covariates parameters should be regularized. (default: TRUE)}

\item{iter_lambda}{Maximum number of iterations for within the lambda analysis. (default: 100)}

\item{seed}{Any integer to control the random seed. NA for a random seed (default: NA)}

\item{file_path}{Path (full or relative) to where the files should be written if write_to_file=true (default: active working directory)}
}
\value{
: Currently no return of the function. Use write_to_file flag to store results to disk.
}
\description{
This function estimate deviations from optimal lambda
If post_estimation is set, then the estimated parameters are returned additionally in the list.
}
\examples{
mcms_lambda_analysis(M, X, Z, B, mask, num_B_cov)
}
