% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing.R
\name{normalize.labelfree}
\alias{normalize.labelfree}
\title{Normalize the sum intensities data and a linear model}
\usage{
normalize.labelfree(data, lm.model)
}
\arguments{
\item{data}{the MS data to normalize}

\item{lm.model}{an optional \code{\link{lm}} model of the form lm(log.Intensity ~ Raw.file). If provided, a column named norm.I will be calculated. Alternative models can be used if they provide compatible methods for \code{\link{coef}} and \code{\link{predict}}}
}
\description{
Normalize the sum intensities data and a linear model
}
