% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{nParams}
\alias{nParams}
\alias{nParams,Peptides-method}
\title{Number of parameters}
\usage{
nParams(object)

\S4method{nParams}{Peptides}(object)
}
\arguments{
\item{object}{the \code{Peptides} model}
}
\value{
an integer
}
\description{
Number of parameters
}
\section{Methods (by class)}{
\itemize{
\item \code{Peptides}: This corresponds to the sum of the concentration and occupancy parameters
}}

\examples{
data(ENSTest, var.model)
ENSTestProtein <- Protein(ENSTest)
ENSTestModel <- Peptides(ENSTestProtein)
nParams(ENSTestModel)
}
