% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptidesClass.R
\docType{class}
\name{Peptides-class}
\alias{Peptides-class}
\title{Set of peptides}
\description{
A class that describe a set of peptides. Use the \code{\link{Peptides}} function for easy object creation
}
\section{Slots}{

\describe{
\item{\code{c}}{the concentration ratios (per sample pair) as a named numeric (name is sampleX_sampleY, ...)}

\item{\code{o}}{the occupancy ratios (per sample) as a named numeric (name is sampleX, ...)}

\item{\code{num.c,num.o}}{number of o and c parameters}

\item{\code{names.c,names.o}}{names of the pair to which each c, and of the sample to which each o applies.}

\item{\code{protein}}{the \code{\link{Protein-class}} object}
}}

