% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.Peptides}
\alias{predict.Peptides}
\title{Predicts ratios from a Peptides model}
\usage{
\method{predict}{Peptides}(object, newdata = object@protein, ...)
}
\arguments{
\item{object}{the \code{\link{Peptides}} model to predict from}

\item{newdata}{an optional \code{\link{Protein}} model to use as data for the predictions}

\item{...}{ignored.}
}
\description{
Predicts ratios from a Peptides model
}
\examples{
data(ENSTest)
ENSTestProtein <- Protein(ENSTest)
ENSTestModel <- Peptides(ENSTestProtein)
predict(ENSTestModel)
}
