% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinClass.R
\docType{class}
\name{Protein-class}
\alias{Protein-class}
\title{A representation of a protein}
\description{
A class that describe a protein. Use the \code{\link{Protein}} function for easy object creation.
}
\section{Slots}{

\describe{
\item{\code{data}}{a data.frame with the data for one protein.}

\item{\code{modifications}}{vector listing all the modification sites}

\item{\code{samples}}{vector with the name of all the samples (unique and sorted)}

\item{\code{references}}{vector with the name of the reference sample(s)}

\item{\code{reference.sample.overlap}}{those sample that appear both as reference and as sample so one can "close the loop"}

\item{\code{reference.sample.intersect}}{all the sample and reference names. Will for instance correspond to all the occupancy ratios that need to be computed.}

\item{\code{sample.dependency}}{a matrix with all the sample_reference pairs in rows, and the columns over all non-redundant pairs. Useful to compute concentration ratios}

\item{\code{sites.activation}}{indicator matrix of which sites are on or off. Meaningful only on the indices where \code{site.coverage} is 1}

\item{\code{sites.coverage}}{indicator matrix of which peptide covers which site}
}}

