% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinClass.R
\name{Protein}
\alias{Protein}
\title{End-user function to create a Protein object}
\usage{
Protein(data, sample.dependency.matrix,
  decimate.sample.dependency.matrix = TRUE, na.action = na.warn)
}
\arguments{
\item{data}{the \code{data.frame}}

\item{sample.dependency.matrix}{a matrix encoding how samples (rows) depend on the parameters (columns). It can be omitted for simple problems with no complex dependencies.}

\item{decimate.sample.dependency.matrix}{whether to remove samples that are not observed and the parameters that are not}

\item{na.action}{either \code{\link{na.warn}}, \code{\link{na.fail}} or \code{\link{na.omit}}. Actions keeping NAs may compromise the sampling.}
}
\description{
Only acetylation (a) is supported as a non-positional modification. It is added to the N-term of the peptide
}
\examples{
data(ENSTest)
Protein(ENSTest)
}
