% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing.R
\name{read.labelfree}
\alias{read.labelfree}
\title{Read a label-free MaxQuant project
This function reads, normalises and calculates ratios from a MaxQuant project ("txt" result folder)}
\usage{
read.labelfree(dir, reference.experiment, score.threshold = 40,
  mod.threshold = 0.9, raw.file.filter = NULL, plot = FALSE)
}
\arguments{
\item{dir}{the \dQuote{txt} folder containing the MaxQuant results}

\item{reference.experiment}{the name of the reference experiment}

\item{score.threshold}{minimal identification score. Peptides with values below this threshold will be filtered out.}

\item{mod.threshold}{minimal site localisation probability accepted. Peptides with values below this threshold will be filtered out.}

\item{raw.file.filter}{an optional filter for the "Raw file" column passed to \code{\link{str_detect}}}

\item{plot}{show diagnostic plots}
}
\description{
Read a label-free MaxQuant project
This function reads, normalises and calculates ratios from a MaxQuant project ("txt" result folder)
}
\examples{
\dontrun{
labelfree <- read.labelfree("searched/txt", reference.experiment = "control", plot = TRUE)
}
}
