% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing_functions.R
\name{read.MQ}
\alias{read.MQ}
\title{Reads a MQ file with the proper format: header, tab separator, no stringsAsFactor, no quotes}
\usage{
read.MQ(MQ.file, ...)
}
\arguments{
\item{MQ.file}{the file name}

\item{...}{additional arguments for \code{\link{read.table}} such as \code{nrow}, etc.}
}
\value{
the MQ file as a \code{\link{data.frame}}
}
\description{
Reads a MQ file with the proper format: header, tab separator, no stringsAsFactor, no quotes
}
\examples{
\dontrun{
evidence <- read.MQ("search/txt/evidence.txt")
}
}
