% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{variance.model}
\alias{variance.model}
\title{Variance model for MS data
This function creates a model of the variance based on the global proteomics data.
The data is cut in bins of similar |log(ratio)| where the variance is asseased. The binning is controlled by the \code{n.cuts}, \code{counts.pr.cut}, \code{cutVector} and \code{cut.type} parameters}
\usage{
variance.model(
  data,
  mean = "ratio",
  q = "q",
  n = "n",
  n.cuts = NULL,
  plot = FALSE,
  verbose = FALSE,
  fixed.shape = FALSE,
  fixed.rate = FALSE,
  min.n = 3,
  alpha = c(0, 0),
  counts.pr.cut = 400,
  cutVector = NULL,
  cut.type = c("quantile", "equal"),
  weight.type = c("normal", "none", "log"),
  n.quantiles.pr.cut = NULL,
  counts.pr.quantile.pr.cut = 20,
  alpha.pr.cut = 0,
  min.variance = 0.005,
  probs.pr.cut = NULL,
  moment.fit = 0,
  rate.from.mean = FALSE,
  weighted.mean.shape = TRUE,
  max.rate = 25,
  ...
)
}
\arguments{
\item{data}{the normalized data}

\item{mean}{the name of the column with the mean ratio}

\item{q}{the name of the column with q = (n-1) * sd^2}

\item{n}{the name of the column with the number of effective observations}

\item{n.cuts}{is the number of cuts along the x=|log(ratio)| axis to estimate conditional shape and rate parameters for the inverse variances}

\item{plot}{whether to show diagnostic plots}

\item{verbose}{whether to print diagnostic messages along the way}

\item{fixed.shape, fixed.rate}{whether to keep shape and rate fixed}

\item{min.n}{use only observations with a number of effective observations greater or equal to that}

\item{alpha}{proportion of the tails of the quantiles to ignore}

\item{counts.pr.cut}{the target number of ratios in each bin}

\item{cutVector}{the direct vector of cut points. Overrides the other binning arguments (\code{n.cuts}, \code{counts.pr.cut} and \code{cut.type})}

\item{cut.type}{whether to cut by \dQuote{quantile} or cuts of \dQuote{equal} size}

\item{weight.type}{type of weighting}

\item{n.quantiles.pr.cut}{number of quantiles for the gamma function}

\item{counts.pr.quantile.pr.cut}{the target number of observation in each quantile for the gamma function}

\item{alpha.pr.cut}{proportion of the tails to ignore in the gamma function}

\item{min.variance}{floor level for the variance}

\item{probs.pr.cut}{probabilities for the gamma function}

\item{moment.fit}{whether to fit the moment. Values of 0, 1 or 2 are accepted}

\item{rate.from.mean}{whether to use the rate in the mean fit}

\item{weighted.mean.shape}{whether to use weighting to calculate the mean shape}

\item{max.rate}{maximum allowed value for the rate}

\item{...}{additional arguments to \code{\link{plot}} and the gamma function}
}
\description{
Variance model for MS data
This function creates a model of the variance based on the global proteomics data.
The data is cut in bins of similar |log(ratio)| where the variance is asseased. The binning is controlled by the \code{n.cuts}, \code{counts.pr.cut}, \code{cutVector} and \code{cut.type} parameters
}
\author{
Jesper Ferkinghoff-Borg
}
